Started django app
This is supposed to be a training in Django.
The first goal is to create an api for bot status. 3 separate apis for
reporting current bot status to be used in web-based admin page.
This is to eventually replace terminal block logger.

Also later down the road web version of the tg-bot may be implemented.