Stable Version 2.9
This commit is assosiated with the start of migration towards volume-less
operation. This is due to increased costs and complexity of the program
when used with separate volume while running in the cloud.

Further development will be aiming towards intercontainer file sharing
via http requests instead of having access to a common volume.

One new container will be introduced to the docker-compose for storage
purposes. This container would store all the files generated by other
containers and erase them on expiry. All other containers would
POST their generated files to this Storage container and GET the
files required for their operation from this Storage container.