This branch was created to facilitate direct recording. The tests were performed on OrangePi5Plus with ubuntuGnome desktop and GPU acceleration.
Results were disappoining - the video was laggy and sluggish. Comparable to recording
of the xvfb display on selenium container.

Source code of the experiment is in video_gfx/video_gfx/direct_recording.
Board specific folder is orangepi5plu. To test it you need to run it on the same
exact board and os with "xhost +" command before. Also populate server/assets with some pictures for gfx.